src/backend/requirements.txt
azure.cosmos
azure.communication.email
azure.identity 
python-dotenv
azure-functions
pytest
requests
PyJWT[crypto]>=2.8.0